+++ source/dsp/GranularEngine.h
@@ -7,7 +7,8 @@
 #include "GrainPool.h"
 #include "GrainProcessor.h"
 #include "PositionModulator.h"
-#include "Scheduler.h"
+//#include "Scheduler.h"
+#include "VoiceManager.h"
 #include "framework/bridge/PingPongBuffer.h"
 #include "framework/core/PluginParams.h"
 #include "framework/core/PluginTypes.h"
@@ -29,7 +30,8 @@ namespace particules
         GranularEngine(FromAudio& fa);
         ~GranularEngine() = default;
 
-        void process(AudioBuffer& outputBuffer, AudioPayload* payload, int bufferSize, float* const* outputPtrs,
+        void process(AudioBuffer& outputBuffer, juce::MidiBuffer& midiBuffer, AudioPayload* payload, int bufferSize,
+            float* const* outputPtrs,
             int outputNumChannels, const ParameterSnapshot& ps);
 
         void init(double, int, int);
@@ -41,29 +43,35 @@ namespace particules
         void updateSmoothedParameters() noexcept;
         void setTargetSmoothedValue(const ParameterSnapshot&) noexcept;
 
-        static constexpr uint8_t mMaxEvent = params::maxSpawnsPerBlock;
+        //static constexpr uint8_t mMaxEvent = params::maxSpawnsPerBlock;
 
         // core data members
         const float refreshRate;
         int sampleAccumulator;
         int threshold;
-        std::function<void(const ParameterSnapshot& ps, AudioPayload* payload)> spawnCallback;
+
+        //SpawnGrainCallback spawnCallback;
 
         // core components
         GrainEnvelope envLut;
         PositionModulator posMod;
-        Scheduler scheduler;
+        VoiceManager voiceManager;
         GrainPool pool;
         GrainProcessor grainProcessor;
 
         // processors
-        juce::dsp::Gain<float> gainProcessor;
-        juce::ADSR adsr;
-        juce::ADSR::Parameters adsrParams;
+        //juce::ADSR adsr;
+        //juce::ADSR::Parameters adsrParams;
 
         // smoothed value : we only smooth parameters that change the life cycle of the grains.
         juce::SmoothedValue<float> speedSmooth;
+        juce::SmoothedValue<float> attackSmooth;
+        juce::SmoothedValue<float> decaySmooth;
+        juce::SmoothedValue<float> sustainSmooth;
+        juce::SmoothedValue<float> releaseSmooth;
+        juce::dsp::Gain<float> gainProcessor;
         //juce::SmoothedValue<float> sustainRatioSmooth;
+
         SmoothedParameters smoothedParams;
 
         FromAudio& faudio;