+++ source/framework/core/PluginParams.h
@@ -24,13 +24,13 @@ namespace particules
     namespace params
     {
 
-        inline constexpr int maxMidiVoice = 16;
+        inline constexpr int maxMidiVoice = 8;
         inline constexpr int maxSpawnsPerBlock = 16;
         inline constexpr double maxFileDuration = 600.0; // 10 min
         inline constexpr uint64_t maxFileSize = 4LL * 512LL * 1024LL * 1024LL; // 512MB limits ULL -> Unsigned Long Long : 2GB
         inline constexpr int maxLutSize = 2048;
         inline constexpr int maxActiveGrains = 500;
-        inline constexpr float maxDuration = 10.0f; // seconds
+        //inline constexpr float maxDuration = 10.0f; // seconds
 
         namespace attack
         {