+++ source/utils/struct/SmoothedParameters.h
@@ -6,5 +6,9 @@ namespace particules
     {
         float speed;
         float sustainRatio;
+        float attack;
+        float decay;
+        float sustain;
+        float release;
     };
 }